Re-organize classes for cleaner structure.
